analysisBD/requirements.txt
dj-database-url==0.4.1
Django==3.1.7
djangorestframework==3.13.1
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==6.2.0
neo4j==4.4.5